Move out settings health to dedicated cron job
