+++ search/includes/classes/class-search.php
@@ -88,6 +88,7 @@ class Search {
 	private const UPPER_BOUND_QUERY_DB_FALLBACK_VALUE = 10;
 
 	public $healthcheck;
+	public $settingshealthcheck;
 	public $versioning_cleanup;
 	public $field_count_gauge;
 	public $queue_wait_time;
@@ -146,6 +147,9 @@ class Search {
 
 		// Load health check cron job
 		require_once __DIR__ . '/class-healthjob.php';
+
+		// Load settings health check cron job
+		require_once __DIR__ . '/class-settingshealthjob.php';
 
 		// Load versioning cleanup job
 		require_once __DIR__ . '/class-versioningcleanupjob.php';
@@ -521,6 +525,7 @@ class Search {
 
 	protected function setup_cron_jobs() {
 		$this->healthcheck = new HealthJob( $this );
+		$this->settingshealthcheck = new SettingsHealthJob( $this );
 		$this->versioning_cleanup = new VersioningCleanupJob( $this->indexables, $this->versioning );
 
 		/**
@@ -530,9 +535,11 @@ class Search {
 		 */
 		if ( defined( 'WP_CLI' ) && \WP_CLI ) {
 			add_action( 'wp_loaded', [ $this->healthcheck, 'init' ], 0 );
+			add_action( 'wp_loaded', [ $this->settingshealthcheck, 'init' ], 0 );
 			add_action( 'wp_loaded', [ $this->versioning_cleanup, 'init' ], 0 );
 		} else {
 			add_action( 'admin_init', [ $this->healthcheck, 'init' ], 0 );
+			add_action( 'admin_init', [ $this->settingshealthcheck, 'init' ], 0 );
 			add_action( 'admin_init', [ $this->versioning_cleanup, 'init' ], 0 );
 		}
 	}
